PRR-57 [refactor]: 어드민 API 패키지 분리
